Rename same origin filter file back to its original name
