+++ app/src/main-process/same-origin-filter.ts
